requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-limiter==2.8.0
flask-bcrypt==1.0.1
pydub==0.25.1
python-dotenv==0.19.0
SpeechRecognition==3.8.1
PyJWT==2.3.0